🔨 refactor: auth entity from auth module moved to common module
