+++ auth/handler/user_handler.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/gofiber/fiber/v3"
-	"github.com/metalpoch/olt-blueprint/auth/model"
 	"github.com/metalpoch/olt-blueprint/auth/usecase"
+	"github.com/metalpoch/olt-blueprint/common/model"
 )
 
 type UserHandler struct {
